pyproject.toml
[project]
name = "pennylane-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "optax>=0.2.5",
    "pennylane>=0.41.1",
    "pennylane-catalyst>=0.11.0",
]
